startup/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==3.7.2